POC_code/requirements.txt
python-bitcoinlib
simplejson
pyzmq
pycrypto
